% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_counter}
\alias{cluster_counter}
\title{Counts the number of observations per cluster}
\usage{
cluster_counter(cluster_location, n_clusters)
}
\arguments{
\item{cluster_location}{a vector with positive intergers containing each
point observation's cluster.}

\item{n_clusters}{total number of clusters}
}
\value{
Vector with number of observations per cluster
}
\description{
Counts the number of observations per cluster
}
