% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cluster_location}
\alias{cluster_location}
\title{Returns each observation's cluster location based on the euclidean distance
between the observation and cluster centers.}
\usage{
cluster_location(x, centers)
}
\arguments{
\item{x}{a numeric matrix with observations.}

\item{centers}{a numeric matrix with observations with cluster's centers
coordinates.}
}
\value{
A list with the following components:
\item{clusters }{a vector indicating each point observation's cluster. }
\item{distance }{a vector with the distance from each
observation to its corresponding cluster center. }
}
\description{
Returns each observation's cluster location based on the euclidean distance
between the observation and cluster centers.
}
