% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{const_c2}
\alias{const_c2}
\title{Tuning constant for \eqn{\tau}-estimator}
\usage{
const_c2(p)
}
\arguments{
\item{p}{dimension where observation lives.}
}
\value{
Tuning constant for \eqn{\tau}-estimator.
}
\description{
Tuning constant approximation for \eqn{\tau}-estimator to reach 90%
efficiency
}
\references{
Maronna, R. A. and Yohai, V. J. (2017). Robust and efficient estimation of
multivariate scatter and location. Computational Statistics & Data Analysis,
109:64–75.
}
