% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derpsi_opt}
\alias{derpsi_opt}
\title{Second derivative of the quasi \eqn{\rho} function}
\usage{
derpsi_opt(x, c)
}
\arguments{
\item{x}{numeric vector with positive values.}

\item{c}{tunning constant.}
}
\value{
Numeric vector with the second derivative of the quasi optimal \eqn{\rho}
computation for each element of x.
}
\description{
Second derivative of the quasi \eqn{\rho} function
}
