% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoOpt.R
\name{derpsiOpt}
\alias{derpsiOpt}
\title{derpsiOpt
the derivative of the psi function  (see \code{\link{psiOpt}})}
\usage{
derpsiOpt(x, cc)
}
\arguments{
\item{x}{A real number.}

\item{cc}{a tunning constant.}
}
\value{
rho'(\code{x} / \code{c}\code{c})
}
\description{
derpsiOpt
the derivative of the psi function  (see \code{\link{psiOpt}})
}
\examples{
psiOpt(x=0.5,cc=1)

}
