% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_to_kNN}
\alias{dist_to_kNN}
\title{Find the k Nearest Neighbors}
\usage{
dist_to_kNN(D, k)
}
\arguments{
\item{D}{a distance matrix, which contains the distances between the rows of
a matrix.}

\item{k}{number of neighbors to find.}
}
\value{
A list with the following components:
\item{dist }{a matrix with distances. }
\item{id }{a matrix with indices of the k nearest neighbors. }
\item{k }{number neighbors used. }
}
\description{
\strong{Ties}: If the kth and the (k+1)th nearest neighbor are tied, then
the neighbor found first is returned and the other one is ignored.
}
