% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.distance}
\alias{.distance}
\title{Distance Matrix Computation}
\usage{
.distance(x)
}
\arguments{
\item{x}{a numeric matrix.}
}
\value{
A numeric matrix with the distances between the rows of a matrix.
}
\description{
Computes and returns the distance matrix using euclidean distance
measure to compute the distances between the rows of a
data matrix.
}
