% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.flag_outliers}
\alias{.flag_outliers}
\title{Flag outliers}
\usage{
.flag_outliers(cutoff, b, ktau)
}
\arguments{
\item{cutoff}{quantile of chi-square to be used as a threshold for outliers
detection.}

\item{b}{break down point.}

\item{ktau}{ktaucenters results.}
}
\value{
Numeric vector with the weight factor for each observation
}
\description{
Flag outliers
}
