% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktaucenters.R
\name{ktaucenters}
\alias{ktaucenters}
\title{ktaucenters}
\usage{
ktaucenters(
  X,
  K,
  centers = NULL,
  tolmin = 1e-06,
  NiterMax = 100,
  nstart = 1,
  startWithKmeans = TRUE,
  startWithROBINPD = TRUE,
  cutoff = 0.999
)
}
\arguments{
\item{X}{numeric matrix  of size n x p.}

\item{K}{the number of cluster.}

\item{centers}{a matrix of size K x p containing the K initial centers,
one at each matrix-row. If centers is NULL a random set of (distinct) rows in  \code{X}
are chosen as the initial centres.}

\item{tolmin}{a tolerance parameter used for the algorithm stopping rule}

\item{NiterMax}{a maximun number of iterations used for the algorithm stopping rule}

\item{nstart}{the number of trials that the base algorithm ktaucenters_aux is run.
If it is greater than 1 and center is not set as NULL, a random set of (distinct) rows
in \code{X} will be chosen as the initial centres.}

\item{startWithKmeans}{TRUE if kmean centers values is included as starting point.}

\item{startWithROBINPD}{TRUE if ROBINDEN estimator is included as starting point}

\item{cutoff}{optional argument for outliers detection - quantiles of chi-square to be used as a threshold for outliers detection, defaults to 0.999}
}
\value{
A list including the estimated K centers and labels for the observations
\itemize{
 \item{\code{centers}}{:   matrix of size K x p, with the estimated K centers.}
 \item{\code{cluster}}{: array of size n x 1  integers labels between 1 and K.}
 \item{\code{tauPath}}{: sequence of tau scale values at each iterations.}
 \item{\code{Wni}}{: numeric array of size n x 1 indicating the weights associated to each observation.}
 \item{\code{emptyClusterFlag}}{: a boolean value. True means that in some iteration there were clusters totally empty}
 \item{\code{niter}}{: number of iterations untill convergence is achived or maximun number of iteration is reached}
 \item{\code{di}}{: distance of each observation to its assigned cluster-center}
 \item{\code{outliers}}{: indices observation that can be considered as outliers}
}
}
\description{
Robust Clustering algorithm based on centers, a robust and efficient version of KMeans.
}
\references{
Gonzalez, J. D., Yohai, V. J., & Zamar, R. H. (2019).
Robust Clustering Using Tau-Scales. arXiv preprint arXiv:1906.08198.
}
