% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_screw.R
\docType{data}
\name{mars_screw}
\alias{mars_screw}
\title{Intensity and saturation values of a picture from mars.}
\format{A list containing information about  pixels of a picture form mars
mainly containing red sand and  metal form Rover itself. List include

\itemize{
\item SI_matrix: A matrix with 5063 rows and 128 columns.
Elements 1 to 64 of each row indicate the Saturation values of pixels in a square cell 8 x 8
whereas elements 65 to 128 of each row indicate the cell's Intensity  values.
\item  An integer matrix of dimension 5063 x 2, each row indicates
 each square cell's locations (x-axis y-axis) at the picture.
}}
\source{
\url{https://www.nasa.gov/mission_pages/msl/multimedia/pia16225.html}
}
\usage{
mars_screw
}
\description{
A dataset containing the Intensity and Saturation values of a picture from Mars
taken from Rover Curiosity.
}
\examples{

# Generate Sintetic data (three cluster well separated)
A <- mars_screw$SI_matrix;
B <- mars_screw$geographic_matrix
## looking for the brightest cells
maximun_at_each_cell<- apply(A[ ,1:64], 1, max)
ten_brightest_cells <-order(maximun_at_each_cell, decreasing=TRUE)[1:10]

## plot locations where the ten brightest cells are.
plot(B, pch=19 )
points(B[ten_brightest_cells, ], pch=19,col="yellow" )


}
\keyword{datasets}
