% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_tolerance}
\alias{max_tolerance}
\title{Estimates maximum tolerance between two given matrix row wise}
\usage{
max_tolerance(x, y)
}
\arguments{
\item{x}{numeric matrix.}

\item{y}{numeric matrix.}
}
\value{
max euclidean difference two given matrix row wise.
}
\description{
Estimates maximum tolerance between two given matrix row wise
}
