% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_cpp}
\alias{median_cpp}
\title{Computes the median of a numeric vector}
\usage{
median_cpp(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
median of a numeric vector.
}
\description{
Computes the median of a numeric vector
}
