% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{point_density}
\alias{point_density}
\title{Estimates the local points density.}
\usage{
point_density(D, k)
}
\arguments{
\item{D}{a distance matrix, which contains the distances between the rows of
a matrix.}

\item{k}{number of neighbors to calculate local point density.}
}
\value{
A vector containing the density values for each point.
}
\description{
Estimates the local points density.
}
