% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_opt}
\alias{psi_opt}
\title{Derivative of the quasi optimal \eqn{\rho} function}
\usage{
psi_opt(x, c)
}
\arguments{
\item{x}{numeric vector with positive values.}

\item{c}{tunning constant.}
}
\value{
Numeric vector with the derivative of the quasi optimal \eqn{\rho}
computation for each element of x.
}
\description{
Derivative of the quasi optimal \eqn{\rho} function
}
