% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoOpt.R
\name{psiOpt}
\alias{psiOpt}
\title{psiOpt
psi function  is the derivative of optimal rho function (see \code{\link{rhoOpt}})}
\usage{
psiOpt(x, cc)
}
\arguments{
\item{x}{A real number.}

\item{cc}{a tunning constant.}
}
\value{
rho''(\code{x}/ \code{c}\code{c})
}
\description{
psiOpt
psi function  is the derivative of optimal rho function (see \code{\link{rhoOpt}})
}
\examples{
psiOpt(x=0.5,cc=1)
}
