% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{robin_center}
\alias{robin_center}
\title{Utility function to estimate robinden center}
\usage{
robin_center(idp, indices, crit_robin)
}
\arguments{
\item{idp}{a vector with containing the inverse density each point.}

\item{indices}{vector with sorted indices.}

\item{crit_robin}{critical robin value.}
}
\value{
Index of the cluster center
}
\description{
Utility function to estimate robinden center
}
