% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tau_scale}
\alias{tau_scale}
\title{\eqn{\tau} scale}
\usage{
tau_scale(u, c, s)
}
\arguments{
\item{u}{numeric vector with positive values.}

\item{c}{tuning constant to regulate estimator's efficiency.}

\item{s}{M scale for the same univariate sample (u).}
}
\value{
\eqn{\tau} scale value
}
\description{
The \eqn{\tau} scale of an univariate sample.
}
