% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weight_factor}
\alias{weight_factor}
\title{Weight factor}
\usage{
weight_factor(wni, clusters)
}
\arguments{
\item{wni}{numeric vector with the weight for each observation.}

\item{clusters}{integer vector with the cluster location for each
observation.}
}
\value{
Numeric vector with the weight factor for each observation
}
\description{
Weight factor
}
