% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derpsiOpt}
\alias{derpsiOpt}
\title{Second derivative of the quasi \eqn{\rho} function}
\usage{
derpsiOpt(x, cc)
}
\arguments{
\item{x}{numeric vector with positive values.}

\item{cc}{tunning constant.}
}
\value{
Numeric vector with the second derivative of the quasi optimal \eqn{\rho}
computation for each element of x.
}
\description{
Second derivative of the quasi \eqn{\rho} function
}
