% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktaucenters.R
\name{ktaucenters}
\alias{ktaucenters}
\title{ktaucenters}
\usage{
ktaucenters(
  x,
  centers,
  nstart = 1L,
  use_kmeans = TRUE,
  use_robin = TRUE,
  max_iter = 100L,
  max_tol = 1e-06,
  cutoff = 0.999
)
}
\arguments{
\item{x}{numeric matrix of size n x p, or an object that can be coerced to a matrix (such as a numeric vector or a data frame with all numeric columns).}

\item{centers}{either the number of clusters, say *k*, or a matrix of initial (distinct) cluster centers. If a number, a random set of (distinct) rows in x is chosen as the initial centers.}

\item{nstart}{if centers is a number, how many random sets should be chosen?}

\item{use_kmeans}{use kmeans centers as starting point?}

\item{use_robin}{use robin algorithm centers as starting point?}

\item{max_iter}{the maximum number of iterations allowed.}

\item{max_tol}{maximum tolerance parameter used for the algorithm as stopping rule.}

\item{cutoff}{quantile of chi-square distribution to be used as a threshold for outliers detection, defaults to 0.999}
}
\value{
A list including the estimated k centers and labels for the observations
\itemize{
 \item{\code{centers}}{:   matrix of size K x p, with the estimated K centers.}
 \item{\code{cluster}}{: array of size n x 1  integers labels between 1 and K.}
 \item{\code{tauPath}}{: sequence of tau scale values at each iterations.}
 \item{\code{Wni}}{: numeric array of size n x 1 indicating the weights
associated to each observation.}
 \item{\code{emptyClusterFlag}}{: a boolean value. True means that in some
iteration there were clusters totally empty}
 \item{\code{niter}}{: number of iterations until convergence is achieved
or maximun number of iteration is reached}
 \item{\code{di}}{: distance of each observation to its assigned cluster-center}
 \item{\code{outliers}}{: indices observation that can be considered as outliers}
}
}
\description{
Robust Clustering algorithm.
}
\references{
Gonzalez, J. D., Yohai, V. J., & Zamar, R. H. (2019).
Robust Clustering Using Tau-Scales. arXiv preprint arXiv:1906.08198.
}
