% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_screw.R
\docType{data}
\name{mars_screw}
\alias{mars_screw}
\title{Intensity and saturation values of a picture from mars.}
\format{
A list containing information about  pixels of a picture form mars mainly
containing red sand and  metal form Rover itself. List include:

\itemize{
\item SI_matrix: A matrix with 5063 rows and 128 columns.
Elements 1 to 64 of each row indicate the Saturation values of pixels in a square cell 8 x 8
whereas elements 65 to 128 of each row indicate the cell's Intensity  values.
\item geographic_matrix: An integer matrix of dimension 5063 x 2, each row indicates
 each square cell's locations (x-axis y-axis) at the picture.
 \item screw_index: the index corresponding to the screw observation (screw_index=4180)
}
}
\source{
\url{https://www.nasa.gov/wp-content/uploads/2023/03/694811main_pia16225-43_full.jpg}
}
\usage{
mars_screw
}
\description{
A dataset containing the Intensity and Saturation values of a picture from Mars
taken from Rover Curiosity.
}
\keyword{datasets}
