% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mscale}
\alias{mscale}
\title{M scale}
\usage{
mscale(u, c, b)
}
\arguments{
\item{u}{numeric vector with positive values.}

\item{c}{a tuning constant. If consistency to standard normal distribution
is desired use \code{\link{normal_consistency_constants}}}

\item{b}{the desired break down point.}
}
\value{
M scale value
}
\description{
The M scale of an univariate sample.
}
