% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscale.R
\name{Mscale}
\alias{Mscale}
\title{Mscale
the M scale of an univariate sample (see reference below)}
\usage{
Mscale(u, b = 0.5, c)
}
\arguments{
\item{u}{an univariate sample of size n.}

\item{b}{the desired break down point}

\item{c}{a tuning constant, if consistency to standard normal distribution is desired use
\code{\link{normal_consistency_constants}}}
}
\value{
the Mscale value
}
\description{
Mscale
the M scale of an univariate sample (see reference below)
}
\examples{
Mscale(u=rnorm(100),c=1)

}
\references{
Maronna, R. A., Martin, R. D., Yohai, V. J., & Salibian-Barrera, M. (2018).
Robust statistics: theory and methods (with R). Wiley.
}
