% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mscale}
\alias{Mscale}
\title{M scale}
\usage{
Mscale(u, c, b)
}
\arguments{
\item{u}{numeric vector with positive values.}

\item{c}{a tuning constant. If consistency to standard normal distribution
is desired use \code{\link{normal_consistency_constants}}.}

\item{b}{the desired break down point.}
}
\value{
M scale value.
}
\description{
The M scale of an univariate sample.
}
\examples{
Mscale(u = rnorm(100), c = 1, b = 0.5)

}
\references{
Maronna, R. A., Martin, R. D., Yohai, V. J., & Salibian-Barrera,
M. (2018). Robust statistics: theory and methods (with R). Wiley.
}
