% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psiOpt}
\alias{psiOpt}
\title{Derivative of the quasi optimal \eqn{\rho} function}
\usage{
psiOpt(x, cc)
}
\arguments{
\item{x}{numeric vector with positive values.}

\item{cc}{tunning constant.}
}
\value{
Numeric vector with the derivative of the quasi optimal \eqn{\rho}
computation for each element of x.
}
\description{
Derivative of the quasi optimal \eqn{\rho} function
}
