% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhoOpt.R
\name{rhoOpt}
\alias{rhoOpt}
\title{rhoOpt function}
\usage{
rhoOpt(x, cc)
}
\arguments{
\item{x}{A real number.}

\item{cc}{a tunning constant.}
}
\value{
rho(\code{x}/\code{c}\code{c})
}
\description{
An implementation of quasi optimal rho functions following reference [1]
}
\examples{
val= rhoOpt(x=0.5,cc=1)
}
\references{
[1] Salibian-Barrera, M., Willems, G., & Zamar, R. (2008).
The fast-tau estimator for regression. Journal of Computational and Graphical Statistics, 17(3), 659-682.
}
